First move the popup preview button and abstract the hover logic
